Updated localization setup to use one single .pot file for all slides in a slideshow package. (The numerous .pot files approach did not fit nicely with Rosetta).
generate-local-slides.sh must now use a brute force method to figure out which slides have been localized. (Pending smarter localization support on the slideshow's end).

Dry run of po2html on default slides, using null.po to talk it into trying, because it insists on reformatting each slide's HTML. This way, formatting is consistent with localized and unlocalized slides for the cmp command.

New build dependency: htmlclean, to optimize space.

+++ generate-local-slides.sh
@@ -10,33 +10,50 @@ SOURCE=.
 BUILD=build
 	
 PODIR=$SOURCE/po/$distro
-SOURCESLIDES=$SOURCE/slides/$distro
 BUILDSLIDES=$BUILD/$distro/slides
 
 echo "directory = new Object();" >> $BUILDSLIDES/directory.js;
 
+for slide in $BUILDSLIDES/*.html; do
+	newslide="$slide.new";
+	mv $slide $newslide;
+	#htmlclean removes all comments and whitespace, leaves behind *.bak
+	htmlclean $newslide && rm "$newslide.bak";
+	#run po2html on source slides for consistent formatting
+	po2html --notidy --progress=none -i po/null.po -t $newslide -o $slide;
+	rm $newslide;
+done;
+
 if which po2html; then
 	for locale in $PODIR/*.po; do
-		localename=$(basename $locale .po);
-		localeslides=$BUILDSLIDES/loc.$localename;
-		[ ! -e $localeslides ] && mkdir -p $localeslides;
-		
-		echo "directory['$localename'] = new Object();" >> $BUILDSLIDES/directory.js;
-		
-		for slide in $SOURCESLIDES/*.html; do
-			slidename=$(basename $slide);
-			outputslide="$localeslides/$slidename";
-			echo "$slide";
-			[ -e $outputslide ] && rm -f $outputslide;
+		if [ -e $locale ]; then
+			localename=$(basename $locale .po);
+			localeslides=$BUILDSLIDES/loc.$localename;
 			
-			po2html --notidy --progress=names -i $locale -t $slide -o $outputslide \
-			&& if [ `diff $slide $outputslide` = ""]; then
-				rm $outputslide;
-				rmdir $localeslides;
-			else
-				echo "directory['$localename']['$slidename'] = true;" >> $BUILDSLIDES/directory.js;
-			fi;
-		done;
+			echo "directory['$localename'] = new Object();" >> $BUILDSLIDES/directory.js;
+			echo "Found locale: $locale";
+			
+			for slide in $BUILDSLIDES/*.html; do
+				slidename=$(basename $slide);
+				[ $slidename = "index.html" ] && continue;
+				
+				outputslide="$localeslides/$slidename";
+				[ -e $outputslide ] && rm -f $outputslide;
+				[ ! -e $localeslides ] && mkdir -p $localeslides;
+				
+				po2html --notidy --progress=names -i $locale -t $slide -o $outputslide;
+				cmp $slide $outputslide;
+				if [ $? -eq 0 ]; then
+					#remove new slide if it's the same as the default
+					rm $outputslide;
+					rmdir $localeslides; #will only succeed if dir is empty
+					echo "$slide was not translated for $locale";
+				else
+					echo "directory['$localename']['$slidename'] = true;" >> $BUILDSLIDES/directory.js;
+					echo "$slide was translated for $locale locale";
+				fi;
+			done;
+		fi;
 	done;
 else
 	echo "\nError: po2html is not available.";
